requirements.txt
customtkinter
packaging
pyinstaller
pywin32
Pillow
pillow-heif
